third_party/blink/web_tests/fast/innerHTML/innerHTML-xml-expected.txt

Test innerHTML in XHTML/XML documents

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS element.firstChild.namespaceURI is "http://www.w3.org/2000/svg"
PASS element.firstChild.namespaceURI is "http://www.w3.org/2000/svg"
PASS element.firstChild.namespaceURI is "http://www.w3.org/2000/svg"
PASS element.firstChild.namespaceURI is "http://www.w3.org/2000/svg"
PASS element.firstChild.namespaceURI is "http://www.w3.org/2000/svg"
PASS element.firstChild.namespaceURI is "http://test/foo"
PASS element.firstChild.namespaceURI is "http://www.w3.org/2000/svg"
PASS element.firstChild.namespaceURI is "http://test/foo"
PASS element.firstChild.namespaceURI is "http://test/foo"
PASS element.firstChild.namespaceURI is "http://www.w3.org/2000/svg"
PASS element.firstChild.namespaceURI is "http://test/foo"
PASS element.firstChild.namespaceURI is "http://www.w3.org/2000/svg"
PASS element.firstChild.namespaceURI is "http://test/foo"
PASS element.firstChild.namespaceURI is "http://test/foo"
PASS element.firstChild.namespaceURI is "http://test/foo"
PASS element.firstChild.namespaceURI is "http://test/foo"
PASS successfullyParsed is true

TEST COMPLETE

 
